// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      lobbies: {
        Row: {
          created_at: string
          description: string | null
          game: Database["public"]["Enums"]["game_type"]
          game_mode: string
          gender_preference: Database["public"]["Enums"]["gender_preference"]
          host_id: string
          id: string
          is_active: boolean
          is_private: boolean
          language: string
          max_players: number
          password: string | null
          region: string | null
          skill_level: Database["public"]["Enums"]["skill_level"]
          title: string
          updated_at: string
          voice_chat: boolean
        }
        Insert: {
          created_at?: string
          description?: string | null
          game: Database["public"]["Enums"]["game_type"]
          game_mode: string
          gender_preference?: Database["public"]["Enums"]["gender_preference"]
          host_id: string
          id?: string
          is_active?: boolean
          is_private?: boolean
          language?: string
          max_players?: number
          password?: string | null
          region?: string | null
          skill_level?: Database["public"]["Enums"]["skill_level"]
          title: string
          updated_at?: string
          voice_chat?: boolean
        }
        Update: {
          created_at?: string
          description?: string | null
          game?: Database["public"]["Enums"]["game_type"]
          game_mode?: string
          gender_preference?: Database["public"]["Enums"]["gender_preference"]
          host_id?: string
          id?: string
          is_active?: boolean
          is_private?: boolean
          language?: string
          max_players?: number
          password?: string | null
          region?: string | null
          skill_level?: Database["public"]["Enums"]["skill_level"]
          title?: string
          updated_at?: string
          voice_chat?: boolean
        }
        Relationships: []
      }
      lobby_members: {
        Row: {
          id: string
          joined_at: string
          lobby_id: string
          user_id: string
        }
        Insert: {
          id?: string
          joined_at?: string
          lobby_id: string
          user_id: string
        }
        Update: {
          id?: string
          joined_at?: string
          lobby_id?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "lobby_members_lobby_id_fkey"
            columns: ["lobby_id"]
            isOneToOne: false
            referencedRelation: "lobbies"
            referencedColumns: ["id"]
          },
        ]
      }
      profiles: {
        Row: {
          avatar_url: string | null
          created_at: string
          game: Database["public"]["Enums"]["game_type"]
          gamer_tag: string
          id: string
          is_verified: boolean | null
          level: number | null
          rank: string | null
          region: string | null
          updated_at: string
          user_id: string
          username: string
        }
        Insert: {
          avatar_url?: string | null
          created_at?: string
          game: Database["public"]["Enums"]["game_type"]
          gamer_tag: string
          id?: string
          is_verified?: boolean | null
          level?: number | null
          rank?: string | null
          region?: string | null
          updated_at?: string
          user_id: string
          username: string
        }
        Update: {
          avatar_url?: string | null
          created_at?: string
          game?: Database["public"]["Enums"]["game_type"]
          gamer_tag?: string
          id?: string
          is_verified?: boolean | null
          level?: number | null
          rank?: string | null
          region?: string | null
          updated_at?: string
          user_id?: string
          username?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      game_type: "cod-mobile" | "pubg-mobile" | "free-fire"
      gender_preference: "any" | "male" | "female" | "mixed"
      skill_level: "beginner" | "intermediate" | "pro" | "any"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never